Relax constraints on dependencies
This is a pretty old crate by Rust standards, and it's forcing the use
of old versions of crates. This causes a problem for our internal builds
because the latest `bindgen` is not compatible with the `1.0` version of
`regex` which this crate requires.

Relaxing these requirements doensn't break any code, and frees us to use
the newest version of `bindgen`